+++ RELEASE_NOTES.md
@@ -1,2 +1,3 @@
 v0.76.1
 
+- Making a chat connection now accepts a locale (Java) or a list of language codes (Swift, TypeScript), which will set the default language to be used for any requests on that connection if provided.